refactor: decompose append.py and effects.py by semantic concern (BTND-07-004)
Resolves CONCERN-2269. Both modules were at the 500-line ceiling, causing
any unrelated edit to require an in-band decomposition.

## status/nodes: append.py → append/ subpackage

- Delete append.py (499 lines, 7 classes)
- Create append/__init__.py — re-exports all 7 public names
- Create append/conditions.py — 4 guard nodes + _has_status_in_participant helper
  (SkipIfIdempotentNode, CheckStatusNotAlreadyAppendedNode,
   ValidateRMTransitionNode, CheckParticipantRMNotClosedNode)
- Create append/actions.py — 3 mutating action nodes
  (LoadParticipantNode, ResolveAndPersistStatusObjectNode,
   AppendStatusAndSaveParticipantNode)
- Mirror test structure: test_append.py → append/conftest.py +
  test_conditions.py + test_actions.py; add TestCheckParticipantRMNotClosedNode
  (was missing from the original test file)

## sync/nodes: effects.py → per-class files + _helpers.py

- Delete effects.py (495 lines, 4 classes + helper)
- Create _helpers.py — _extract_id_from_field + _LedgerEffectNode base class
  (DRY: all 4 effect nodes had identical setup() + _require_log_entry pattern)
- Create participant_status_effect.py, note_effect.py,
  invite_accept_effect.py, close_case_effect.py — one file per class
- Update offer_report_effect.py, ownership_effects.py — import
  _extract_id_from_field from _helpers instead of deleted effects.py
- Update sync/nodes/__init__.py — 4 separate per-class imports
- Mirror test structure: test_effects.py → test_participant_status_effect.py +
  test_note_effect.py + test_invite_accept_effect.py + test_close_case_effect.py;
  adds tests for the 3 previously untested effect nodes
- Update notes/structured-logging.md — reference updated to close_case_effect.py

Note: 5 other modules are approaching the 500-line ceiling
(replay.py 498, suggest_actor/emit.py 498, deploy_fix.py 497,
embargo/nodes/lifecycle.py 494, conditions.py 488) — tracked as future work.
