+++ test/core/behaviors/sync/nodes/test_participant_status_effect.py
@@ -1,5 +1,5 @@
 #!/usr/bin/env python
-"""Regression tests for ApplyParticipantStatusFromLedgerNode (effects.py).
+"""Tests for ApplyParticipantStatusFromLedgerNode.
 
 Covers the critical round-trip serialization bug: a CORE ParticipantStatus
 appended directly to as_CaseParticipant.participant_statuses was serialized with
@@ -25,19 +25,16 @@ from test.core.behaviors.sync.nodes.conftest import (
     _make_event,
     _to_persistable_entry,
 )
-from vultron.adapters.driven.datalayer_sqlite import SqliteDataLayer
-from vultron.core.behaviors.bridge import BTBridge
-from vultron.core.behaviors.sync.nodes.effects import (
+from vultron.core.behaviors.sync.nodes.participant_status_effect import (
     ApplyParticipantStatusFromLedgerNode,
 )
-from vultron.core.models.case_actor import VultronCaseActor
 from vultron.core.models.case_ledger import (
     compute_genesis_hash,
     HashChainLedgerRecord,
 )
+from vultron.core.models.participant_status import ParticipantStatus
 from vultron.core.states.cs import CS_vfd
 from vultron.core.states.rm import RM
-from vultron.core.models.participant_status import ParticipantStatus
 from vultron.wire.as2.vocab.objects.case_participant import as_CaseParticipant
 
 _FIXED_CREATED_AT = datetime(2024, 1, 1, 0, 0, 0, tzinfo=timezone.utc)
@@ -66,11 +63,6 @@ def _make_participant_status_snapshot(
     vfd_state: str = "VFd",
     rm_state: str = "ACCEPTED",
 ) -> dict:
-    """Return a payload_snapshot dict as produced by build_activity_payload_snapshot.
-
-    Uses camelCase keys (wire/alias format) matching how the Case Actor builds
-    the snapshot from an Add(ParticipantStatus, as_CaseParticipant) activity.
-    """
     return {
         "object": {
             "id": status_id,
@@ -91,7 +83,6 @@ def _make_status_entry(
     vfd_state: str = "VFd",
     rm_state: str = "ACCEPTED",
 ):
-    """Return a VultronCaseLedgerEntry for an add_participant_status_to_participant event."""
     snapshot = _make_participant_status_snapshot(
         status_id=status_id,
         participant_id=participant_id,
@@ -111,27 +102,6 @@ def _make_status_entry(
 
 
 @pytest.fixture
-def datalayer():
-    return SqliteDataLayer("sqlite:///:memory:")
-
-
-@pytest.fixture
-def bridge(datalayer):
-    return BTBridge(datalayer=datalayer)
-
-
-@pytest.fixture
-def case_actor(datalayer):
-    actor = VultronCaseActor(
-        name="Case Actor",
-        attributed_to=OWNER_ACTOR_ID,
-        context=CASE_ID,
-    )
-    datalayer.create(actor)
-    return actor
-
-
-@pytest.fixture
 def participant(datalayer):
     p = _make_participant()
     datalayer.save(p)
@@ -147,15 +117,9 @@ def test_apply_participant_status_roundtrip_preserves_vfd_state(
     was serialized with default values (vfd_state='vfd') rather than actual
     values.  After the fix the saved participant must have the correct vfd_state
     from the ledger entry payload snapshot.
-
-    as_CaseParticipant always auto-creates one default ParticipantStatus
-    (RM.START, CS_vfd.vfd) on construction.  After applying the ledger entry,
-    the participant has the initial default PLUS the new status.  The
-    regression manifests as the new status carrying default vfd/rm values
-    instead of the values from the snapshot.
     """
     status_id = f"urn:uuid:{uuid.uuid4()}"
-    initial_count = len(participant.participant_statuses)  # always ≥ 1
+    initial_count = len(participant.participant_statuses)
 
     entry = _make_status_entry(
         status_id=status_id,
@@ -176,25 +140,12 @@ def test_apply_participant_status_roundtrip_preserves_vfd_state(
     assert result.status == Status.SUCCESS
 
     updated = cast(as_CaseParticipant, datalayer.read(participant.id_))
-    assert (
-        updated is not None
-    ), "Participant must still be readable after status update"
-    assert len(updated.participant_statuses) == initial_count + 1, (
-        f"Participant must have exactly one new status appended"
-        f" (expected {initial_count + 1}, got {len(updated.participant_statuses)})"
-    )
+    assert updated is not None
+    assert len(updated.participant_statuses) == initial_count + 1
 
-    # The last (newest) status in the list must carry the values from the
-    # ledger snapshot, not Pydantic serialization defaults.
     new_status = cast(ParticipantStatus, updated.participant_statuses[-1])
-    assert new_status.vfd.state == CS_vfd.VFd, (
-        f"vfd.state must be VFd, got {new_status.vfd.state!r} — "
-        "likely caused by CORE ParticipantStatus serialization mismatch "
-        "when appended to participant_statuses"
-    )
-    assert (
-        new_status.rm.state == RM.ACCEPTED
-    ), f"rm.state must be ACCEPTED, got {new_status.rm.state!r}"
+    assert new_status.vfd.state == CS_vfd.VFd
+    assert new_status.rm.state == RM.ACCEPTED
 
 
 def test_apply_participant_status_idempotent(
@@ -221,13 +172,11 @@ def test_apply_participant_status_idempotent(
 
     updated = cast(as_CaseParticipant, datalayer.read(participant.id_))
     assert updated is not None
-    assert (
-        len(updated.participant_statuses) == initial_count + 1
-    ), "Idempotent apply must not duplicate status entries"
+    assert len(updated.participant_statuses) == initial_count + 1
 
 
 def test_apply_participant_status_skips_missing_participant(
-    bridge, datalayer, case_actor
+    bridge, case_actor
 ):
     """Node returns SUCCESS without error when participant not found locally."""
     status_id = f"urn:uuid:{uuid.uuid4()}"
