+++ vultron/core/behaviors/status/nodes/append/__init__.py
@@ -1,0 +1,49 @@
+#!/usr/bin/env python
+
+#  Copyright (c) 2026 Carnegie Mellon University and Contributors.
+#  - see Contributors.md for a full list of Contributors
+#  - see ContributionInstructions.md for information on how you can Contribute to this project
+#  Vultron Multiparty Coordinated Vulnerability Disclosure Protocol Prototype is
+#  licensed under a MIT (SEI)-style license, please see LICENSE.md distributed
+#  with this Software or contact permission@sei.cmu.edu for full terms.
+#  Created, in part, with funding and support from the United States Government
+#  (see Acknowledgments file). This program may include and/or can make use of
+#  certain third party source code, object code, documentation and other files
+#  ("Third Party Software"). See LICENSE.md for more details.
+#  Carnegie Mellon®, CERT® and CERT Coordination Center® are registered in the
+#  U.S. Patent and Trademark Office by Carnegie Mellon University
+
+"""Append-participant-status leaf nodes subpackage.
+
+Re-exports all public node classes from the two domain-specific submodules so
+that existing import paths (``from vultron.core.behaviors.status.nodes.append
+import ...``) continue to work without modification.
+
+Submodules:
+- ``conditions``: Idempotency guards and RM-transition validation nodes
+- ``actions``: DataLayer-mutating load, resolve, and append action nodes
+"""
+
+from vultron.core.behaviors.status.nodes.append.actions import (
+    AppendStatusAndSaveParticipantNode,
+    LoadParticipantNode,
+    ResolveAndPersistStatusObjectNode,
+)
+from vultron.core.behaviors.status.nodes.append.conditions import (
+    CheckParticipantRMNotClosedNode,
+    CheckStatusNotAlreadyAppendedNode,
+    SkipIfIdempotentNode,
+    ValidateRMTransitionNode,
+)
+
+__all__ = [
+    # conditions
+    "SkipIfIdempotentNode",
+    "CheckStatusNotAlreadyAppendedNode",
+    "ValidateRMTransitionNode",
+    "CheckParticipantRMNotClosedNode",
+    # actions
+    "LoadParticipantNode",
+    "ResolveAndPersistStatusObjectNode",
+    "AppendStatusAndSaveParticipantNode",
+]
