+++ vultron/core/behaviors/status/nodes/append/conditions.py
@@ -13,31 +13,41 @@
 #  Carnegie Mellon®, CERT® and CERT Coordination Center® are registered in the
 #  U.S. Patent and Trademark Office by Carnegie Mellon University
 
-"""Append-participant-status leaf nodes for DEMOMA-07-003 step 2.
+"""Condition (guard) nodes for the append-participant-status workflow.
 
-Contains the five leaf nodes that implement the append sequence:
-load participant, check idempotency, resolve status object, validate RM
-transition, and append + save.
+Contains the four precondition and idempotency-guard nodes:
+
+- :class:`SkipIfIdempotentNode` — Selector-level skip when status is already
+  present.
+- :class:`CheckStatusNotAlreadyAppendedNode` — Sequence-level guard: halt if
+  already appended.
+- :class:`ValidateRMTransitionNode` — Reject backwards RM transitions.
+- :class:`CheckParticipantRMNotClosedNode` — Pre-flight guard: reject rewrites
+  for already-closed participants.
 """
 
 import logging
-from typing import Any, cast
+from typing import Any
 
 import py_trees
 from py_trees.common import Status
 
-from vultron.core.behaviors.helpers import DataLayerAction, DataLayerCondition
+from vultron.core.behaviors.helpers import DataLayerCondition
 from vultron.core.models.case_participant import CaseParticipant
-from vultron.core.models.participant_status import ParticipantStatus
-from vultron.core.models.protocols import PersistableModel
+from vultron.core.models._helpers import _as_id
 from vultron.core.states.rm import (
     RM,
     is_monotonic_rm_forward,
     is_valid_rm_transition,
 )
-from vultron.core.models._helpers import _as_id
 
 logger = logging.getLogger(__name__)
+
+
+def _has_status_in_participant(participant: Any, status_id: str) -> bool:
+    """Return True when *status_id* is already in the participant's status list."""
+    existing_ids = [_as_id(s) for s in participant.participant_statuses]
+    return status_id in existing_ids
 
 
 class SkipIfIdempotentNode(py_trees.behaviour.Behaviour):
@@ -66,6 +76,7 @@ class SkipIfIdempotentNode(py_trees.behaviour.Behaviour):
         self.participant_id = participant_id
 
     def setup(self, **kwargs: Any) -> None:
+        super().setup(**kwargs)
         self.blackboard = py_trees.blackboard.Client(name=self.name)
         self.blackboard.register_key(
             key="append_status_participant",
@@ -77,8 +88,7 @@ class SkipIfIdempotentNode(py_trees.behaviour.Behaviour):
         if participant is None:
             return Status.FAILURE
 
-        existing_ids = [_as_id(s) for s in participant.participant_statuses]
-        if self.status_id in existing_ids:
+        if _has_status_in_participant(participant, self.status_id):
             logging.getLogger(self.__class__.__module__).info(
                 "SkipIfIdempotentNode: status '%s' already on participant"
                 " '%s' — idempotent, skipping (SUCCESS)",
@@ -87,52 +97,6 @@ class SkipIfIdempotentNode(py_trees.behaviour.Behaviour):
             )
             return Status.SUCCESS
         return Status.FAILURE
-
-
-class LoadParticipantNode(DataLayerAction):
-    """Load the CaseParticipant from DataLayer to blackboard.
-
-    Reads the participant by ID and writes it to the blackboard under the key
-    ``append_status_participant``.
-
-    Returns SUCCESS if the participant is found and is a valid participant model.
-    Returns FAILURE if participant not found or is not a participant model.
-    """
-
-    def __init__(self, participant_id: str, name: str | None = None):
-        super().__init__(name=name or self.__class__.__name__)
-        self.participant_id = participant_id
-
-    def setup(self, **kwargs: Any) -> None:
-        super().setup(**kwargs)
-        self.blackboard.register_key(
-            key="append_status_participant",
-            access=py_trees.common.Access.WRITE,
-        )
-
-    def update(self) -> Status:
-        if (f := self._require_datalayer()) is not None:
-            return f
-        assert self.datalayer is not None
-
-        participant = self.datalayer.read(self.participant_id)
-        if not isinstance(participant, CaseParticipant):
-            self.feedback_message = (
-                f"Participant '{self.participant_id}' not found"
-            )
-            self.logger.warning(
-                "LoadParticipantNode: %s", self.feedback_message
-            )
-            return Status.FAILURE
-
-        self.logger.debug(
-            "LoadParticipantNode: loaded participant '%s'",
-            self.participant_id,
-        )
-        self.blackboard.set(
-            "append_status_participant", participant, overwrite=True
-        )
-        return Status.SUCCESS
 
 
 class CheckStatusNotAlreadyAppendedNode(DataLayerCondition):
@@ -170,8 +134,7 @@ class CheckStatusNotAlreadyAppendedNode(DataLayerCondition):
             )
             return Status.FAILURE
 
-        existing_ids = [_as_id(s) for s in participant.participant_statuses]
-        if self.status_id in existing_ids:
+        if _has_status_in_participant(participant, self.status_id):
             self.logger.info(
                 "CheckStatusNotAlreadyAppendedNode: status '%s' already"
                 " on participant '%s' — idempotent, skipping",
@@ -183,79 +146,6 @@ class CheckStatusNotAlreadyAppendedNode(DataLayerCondition):
         self.logger.debug(
             "CheckStatusNotAlreadyAppendedNode: status '%s' not yet appended",
             self.status_id,
-        )
-        return Status.SUCCESS
-
-
-class ResolveAndPersistStatusObjectNode(DataLayerAction):
-    """Resolve the status object by ID, persisting fallback if needed.
-
-    Tries the DataLayer first; if not found, uses ``status_obj_fallback``,
-    saves it, then re-reads the canonical wire-format record.
-
-    Validates that the resolved object is a ParticipantStatus (has rm and
-    vfd attributes).
-
-    Writes the resolved status object to the blackboard under the key
-    ``append_status_status_obj``.
-
-    Returns SUCCESS if status is resolved and valid.
-    Returns FAILURE if status cannot be resolved or is not a ParticipantStatus.
-    """
-
-    def __init__(
-        self,
-        status_id: str,
-        status_obj_fallback: PersistableModel | None,
-        name: str | None = None,
-    ):
-        super().__init__(name=name or self.__class__.__name__)
-        self.status_id = status_id
-        self.status_obj_fallback = status_obj_fallback
-
-    def setup(self, **kwargs: Any) -> None:
-        super().setup(**kwargs)
-        self.blackboard.register_key(
-            key="append_status_status_obj",
-            access=py_trees.common.Access.WRITE,
-        )
-
-    def update(self) -> Status:
-        if (f := self._require_datalayer()) is not None:
-            return f
-        assert self.datalayer is not None
-
-        status_obj = self.datalayer.read(self.status_id)
-        if not hasattr(status_obj, "id_"):
-            status_obj = self.status_obj_fallback
-            if status_obj is not None:
-                self.datalayer.save(status_obj)
-                status_obj = self.datalayer.read(self.status_id) or status_obj
-
-        if status_obj is None or not hasattr(status_obj, "id_"):
-            self.feedback_message = f"Status '{self.status_id}' not found"
-            self.logger.warning(
-                "ResolveAndPersistStatusObjectNode: %s",
-                self.feedback_message,
-            )
-            return Status.FAILURE
-
-        if not hasattr(status_obj, "rm") or not hasattr(status_obj, "vfd"):
-            self.feedback_message = (
-                f"Object '{self.status_id}' is not a ParticipantStatus"
-            )
-            self.logger.warning(
-                "ResolveAndPersistStatusObjectNode: %s",
-                self.feedback_message,
-            )
-            return Status.FAILURE
-
-        self.logger.debug(
-            "ResolveAndPersistStatusObjectNode: resolved status '%s'",
-            self.status_id,
-        )
-        self.blackboard.set(
-            "append_status_status_obj", status_obj, overwrite=True
         )
         return Status.SUCCESS
 
@@ -362,63 +252,6 @@ class ValidateRMTransitionNode(DataLayerCondition):
         return Status.FAILURE
 
 
-class AppendStatusAndSaveParticipantNode(DataLayerAction):
-    """Append the status object to the participant and save.
-
-    Appends the resolved status object (from blackboard) to the participant's
-    status list and saves the participant to the DataLayer.
-
-    Returns SUCCESS on successful append and save.
-    Returns FAILURE if participant or status not on blackboard.
-    """
-
-    def __init__(
-        self, status_id: str, participant_id: str, name: str | None = None
-    ):
-        super().__init__(name=name or self.__class__.__name__)
-        self.status_id = status_id
-        self.participant_id = participant_id
-
-    def setup(self, **kwargs: Any) -> None:
-        super().setup(**kwargs)
-        self.blackboard.register_key(
-            key="append_status_participant",
-            access=py_trees.common.Access.READ,
-        )
-        self.blackboard.register_key(
-            key="append_status_status_obj",
-            access=py_trees.common.Access.READ,
-        )
-
-    def update(self) -> Status:
-        if (f := self._require_datalayer()) is not None:
-            return f
-        assert self.datalayer is not None
-
-        participant = self.blackboard.get("append_status_participant")
-        status_obj = self.blackboard.get("append_status_status_obj")
-
-        if participant is None or status_obj is None:
-            self.feedback_message = "Participant or status not on blackboard"
-            self.logger.warning(
-                "AppendStatusAndSaveParticipantNode: %s",
-                self.feedback_message,
-            )
-            return Status.FAILURE
-
-        participant.participant_statuses.append(
-            cast(ParticipantStatus, status_obj)
-        )
-        self.datalayer.save(participant)
-        self.logger.info(
-            "AppendStatusAndSaveParticipantNode: added status '%s' to"
-            " participant '%s' (DEMOMA-07-003 step 2)",
-            self.status_id,
-            self.participant_id,
-        )
-        return Status.SUCCESS
-
-
 class CheckParticipantRMNotClosedNode(DataLayerCondition):
     """Pre-flight guard: FAILURE when participant is in RM.CLOSED with no prior
     status match.
@@ -472,20 +305,17 @@ class CheckParticipantRMNotClosedNode(DataLayerCondition):
 
         # Participant is CLOSED. Allow if the incoming status was already
         # appended by the trigger side (idempotent re-delivery of VALID→CLOSED).
-        if self.status_id:
-            existing_ids = [
-                _as_id(s)
-                for s in getattr(participant, "participant_statuses", [])
-            ]
-            if self.status_id in existing_ids:
-                self.logger.debug(
-                    "%s: participant '%s' is CLOSED but status '%s' already"
-                    " in participant_statuses — allowing idempotent commit",
-                    self.name,
-                    self.participant_id,
-                    self.status_id,
-                )
-                return Status.SUCCESS
+        if self.status_id and _has_status_in_participant(
+            participant, self.status_id
+        ):
+            self.logger.debug(
+                "%s: participant '%s' is CLOSED but status '%s' already"
+                " in participant_statuses — allowing idempotent commit",
+                self.name,
+                self.participant_id,
+                self.status_id,
+            )
+            return Status.SUCCESS
 
         self.feedback_message = (
             f"Participant '{self.participant_id}' is already in terminal"
