+++ vultron/core/behaviors/sync/nodes/offer_report_effect.py
@@ -30,7 +30,7 @@ import py_trees
 from py_trees.common import Status
 
 from vultron.core.behaviors.helpers import DataLayerAction
-from vultron.core.behaviors.sync.nodes.effects import _extract_id_from_field
+from vultron.core.behaviors.sync.nodes._helpers import _extract_id_from_field
 
 
 class ApplyOfferReportFromLedgerNode(DataLayerAction):
